Refactor db related files structure
